Moved all new_headers to parent directory, and it all compiles. Also, greatly improved the Makefile
