+++ CentralComputing/Pod_State.h
@@ -6,7 +6,6 @@
 #include <string>
 #include <SafeQueue.hpp>
 #include "Sensor.h"
-#include "Sensor_Package.h"
 
 class Pod_State : public StateMachine {
 	public:
@@ -27,10 +26,8 @@ class Pod_State : public StateMachine {
 		/** 
 		* Constructs a pod state machine
 		**/
-		Pod_State(SafeQueue<Xmega_Command_t> * cmd_queue, Sensor_Package * sen) 
-      : StateMachine(ST_MAX_STATES), 
-        command_queue(cmd_queue),
-        sensors(sen)
+		Pod_State() 
+      : StateMachine(ST_MAX_STATES)
     {}
 
 		// returns the current state as an enum
@@ -89,9 +86,6 @@ class Pod_State : public StateMachine {
 			STATE_MAP_ENTRY(&Pod_State::ST_Flight_Coast)
 			STATE_MAP_ENTRY(&Pod_State::ST_Flight_Brake)
 		END_STATE_MAP
-
-    SafeQueue<Xmega_Command_t> * command_queue;
-    Sensor_Package * sensors;
 
 };
