+++ CentralComputing/Sensor.cpp
